- added new query parameter trigger and branch to the build resource
By calling /api/:owner/:project/builds?trigger=Travis&branch=notmaster it is
possible to mark how trigger the build and to specify an other than the master branch
as was build. The branch will be neede in the feature to link the tests with it source on github.
Has only be specified if the build was not the master. For an example how to
specify it look at the upload_junitreport.sh script of this project.