- restructure unit reports files for testing by type
The unit test reports files from sbt and karma has some little differences
not in xml elements but by the occurence of them. So split the
test unit reports files by type sbt and karma at the moment could also be added
from other sources to test the compatibility with the parsing here.