- finished implementation of build limit feature
The maximum builds per project are restricted by configuration to ten that
means the last ten builds and it test data are stay in the database. That means if
a new build is upload than it will check if they are more than ten builds than they
will be deleted and all it associated test data.

I use a free mysql instance on heroku for that project so to avoid database pollution
i restrict the maximum data per project. That restriction can be globaly deactived
by configuration of buildslimit in application.conf by setting it to a value equal or
less than zero.

See https://trello.com/c/8VocyKNW