- added remove of maybe existing coverage folder created by instanbul
The instrumentation with istanbul create the folder public/js/bankapp/coverage
that contins the instrumented javascript files. This task instrument all
javascript files in the folder public/js/bankapp so ensure that before a
instrumention task is performed the output coverage folder not exists because
instanbul instrumented the already instrumented files in the coverage folder again.
To prevent that the output folder will be deleted before calling istanbul.