- implements end2end test of the prodcution version
Added new test layer to check after an deployment to heroku if the app
response. It will only check if the main pages are available like overview,
builds, testsuites and testcases. The problem with tests here is that
this test are performed against th elive database so on the test scope
there is unknown what kind of data are returned. To minimize the data influence
the unitcover project itself is used to test the pages except the overview page.
This test layer has the highest possible abstraction and also some limitations.
It could be used to test application flow but not validate any kind of data only
presence of unknoe data from a user perspective.